requirements.txt
discord.py
requests
Flask
gunicorn
